modify git operations
