+++ lib/BasicJavaApplicationBuilder.js
@@ -1,7 +1,7 @@
 'use strict';
 const {OpenShiftClientX} = require('@bcgov/pipeline-cli')
 const Jira = require('./Jira')
-const Git = require('../lib/GitVerification')
+const Git = require('../lib/GitOperation')
 
 module.exports = class {
   constructor(settings) {
@@ -19,7 +19,7 @@ module.exports = class {
     const env = 'build'
     const username = settings.phases[env].credentials.idir.user
     const password = settings.phases[env].credentials.idir.pass
-
+    Object.assign(settings.options.git, {credentials: {'username':username,'password':password}})
     const changeBranch = settings.options.git.branch.merge
     const gitUrl = settings.options.git.url
     const gitElements = gitUrl.split('/')
@@ -38,7 +38,7 @@ module.exports = class {
 
       await this._createJiraAutoRFDs(jiraUrl, repoName, changeBranch, branchName, username, password);
 
-      await this._gitTargetSyncVerify(gitUrl, changeBranch, changeTarget, username, password);
+      await this._gitTargetSyncVerify(Object.assign(settings.options.git));
     }
     else if (changeTarget != EMPTY) {
       // Build for target = 'other branch'
@@ -57,9 +57,9 @@ module.exports = class {
   /**
    * Calling Git module to verify if both change/target branches are not out of sync.
    */
-  async _gitTargetSyncVerify(gitUrl, changeBranch, changeTarget, username, password) {
-    const git = new Git();
-    let result = await git.verify(changeBranch, changeTarget, username, password, gitUrl);
+  async _gitTargetSyncVerify() {
+    const git = new Git(this.settings.options.git);
+    let result = await git.verify();
     if(result == 'True') {
       console.log("Successfully Verified that branch is not out of sync with target")
     }
