+++ lib/GitOperation.js
@@ -7,27 +7,42 @@ const path = require('path');
 const url = require('url')
 
 module.exports = class GitOp {
-   constructor(baseDir, gitRepoUrl, gitCredentials, workingBranch) {
-      this.baseDir = baseDir;
-      this.gitRepoUrl = gitRepoUrl;
-      this.gitCredentials = gitCredentials;
-      this.workingBranch = workingBranch;
+   constructor(settings) {
+      this.settings=settings
    }
+    // Get Merge Base
 
+ 
+   async run(args) {
+      // console.log(`GitVerification: run with args ${args}`);
+             return new Promise( (resolve, reject) =>{
+                const child = spawn('git', args)
+                let stdout = ""
+                let stderr = ""
+                child.stderr.on('data', (data) => {
+                   stderr += data;
+                })
+                child.stdout.on('data', (data) => {
+                   stdout += data;
+                });
+                child.on("exit", (exitCode)=>{
+                   if (exitCode == 0){
+                      resolve({stdout, stderr, exitCode});
+                   }else {
+                      console.log(stderr);
+                      reject(`Exit code ${exitCode}`)
+                   }
+                })
+             });
+    }
    /**
     * This will empty and prepare the baseDir diredtory, clone the repo 
     * and checkout the working branch.
     * Return the baseDir when prepare is done.
     */
-   async prepare() {
+/*    async prepare() {
       const baseDir = this.baseDir;
       return Promise.resolve()
-      .then(()=>{
-         if (fs.existsSync(baseDir)){
-            console.warn("Working directory not empty, removing content.")
-            return childProcess('rm', ['-rf', baseDir]);
-         }
-      })
       .then(()=>{
          return childProcess('git',['init', baseDir])
       })
@@ -47,9 +62,9 @@ module.exports = class GitOp {
       .then(()=>{
          return baseDir;
       })
-   }
+   } */
 
-   async _setupCredentials(workDir, remoteUrl, cred) {
+    async _setupCredentials(workDir, remoteUrl, cred) {
       // git config --global credential.helper 'store --file ~/.git-credentials'
       // git credential-store --file ~/.git-credentials store < .gitcredentials
       // result: https://fakeUsername:fakePass@apps.nrs.gov.bc.ca
@@ -76,7 +91,7 @@ module.exports = class GitOp {
     * @param {*} host host to contact
     * @param {*} gitCredentials the credential to checkout from git.
     */
-   async _setupGitGlobalCredential(protocol, host, gitCredentials) {
+/*    async _setupGitGlobalCredential(protocol, host, gitCredentials) {
       // git config --global credential.helper 'store --file ~/.git-credentials'
       // git credential-store --file ~/.git-credentials store < .gitcredentials
       // result: https://fakeUsername:fakePass@apps.nrs.gov.bc.ca
@@ -90,8 +105,55 @@ module.exports = class GitOp {
       });
 
    } // _setupGitGlobalCredential
-
+ */
    async clear() {
       // TODO: remove the credential line from '.gitcredentials', else?
+   }
+   getMergeBase(changeBranch,changeTarget){
+      return this.run(['fetch','origin', changeBranch])
+      .then(()=>{
+         return this.run(['fetch','origin', changeTarget])
+      })
+      .then(()=>{
+         return this.run(['merge-base',`remotes/origin/${changeBranch}`,`remotes/origin/${changeTarget}`])
+      })
+      .then((output) =>  {
+         return output.stdout.split('\n')[0]
+      })
+    } 
+
+   // Get latest Commit
+   getLatestCommitOnTarget(changeTarget){
+      return this.run(['fetch','origin', changeTarget])
+      .then(()=>{
+         return this.run(['rev-parse',`remotes/origin/${changeTarget}`])
+      })
+      .then((output) =>  {
+         return output.stdout.split('\n')[0]
+      })
+   }
+    // Verify that latest commit is equal to merge-base commit
+   async verify(){
+      await this._setupCredentials(this.settings.dir,this.settings.url,this.settings.credentials)
+               
+            return this.getMergeBase(this.settings.branch.name,this.settings.change.target).then((commitHash1) =>  {
+                  return this.getLatestCommitOnTarget(this.settings.change.target).then((commitHash2) =>{
+                     return new Promise( (resolve) =>{
+                        console.log(`Source Branch Common Ancestor commit: ${commitHash1}`);
+                        console.log(`Target Branch Last Commit - commitHash2: ${commitHash2}`);
+                        if(commitHash1==commitHash2){
+                           resolve('True')
+                        }
+                        else{
+                           throw new Error("\n --------------------------------------- \n Status: Cannot be Merged! \n Reason: Your branch is out of sync from target. \n Solution: Rebase, push and then run the Pipeline \n  ---------------------------------------")
+                        }
+                     });
+                  });
+               });
+
+       //const basedir=__dirname
+       //const git=new GitOps(basedir,url,credentials,changeBranch)
+       //git.prepare()
+      
    }
 } //end GitOp class
